refactor: minor refactor according to review
