+++ src/components/App/config/pages.js
