+++ src/components/History/HistoryFunctional/index.jsx
@@ -5,7 +5,7 @@ import { useSelector } from 'react-redux';
 import { EMPTY_HISTORY_TITLE, HISTORY_TITLE } from '@/constants';
 import historySelector from '@/redux/selectors/historySelector';
 
-import { Border, HistoryItem, HistoryItemWrapper, HistoryWrapper, ItemsWrapper, Title, TitleWrapper } from './styles';
+import { HistoryItem, HistoryItemWrapper, HistoryWrapper, ItemsWrapper, Title, TitleWrapper } from './styles';
 
 const History = () => {
     const { history, showHistory } = useSelector(historySelector);
@@ -13,7 +13,6 @@ const History = () => {
 
     return (
       <React.Fragment>
-        <Border />
         <HistoryWrapper showHistory={showHistory}>
           <TitleWrapper>
             <Title>{history.length ? t(HISTORY_TITLE) : t(EMPTY_HISTORY_TITLE)}</Title>
